refactor: 상품 조회 기능 개선 및 브랜드 정보 포함

+++ apps/commerce-api/src/main/java/com/loopers/application/product/ProductResult.java
@@ -11,7 +11,7 @@ public record ProductResult(
         String description,
         BigDecimal price,
         int stock,
-        Long likeCount,
+        int likeCount,
         Long brandId,
         String brandName,
         String brandDescription
